Done
Needs refactoring, but it's done so far. Should figure out how to use ternaries to display only the buttons needed in a single Student.js Element, but I can't figure out how to do that without importing all the data into the Student.js Element. (Then doing something to the effect of "if student in BC crew return only the binary button")